Ticker: Store previous period length in variable and add getter
This will be useful for determining if frequencies are low enough or high enough and calculating the efficiency of methods.  If a thing doesn't work because the Thread.sleep call is given a negative number (i.e. the tick() function takes longer than the prescribed period length), we can detect this more easily using this and remedy it by externally bumping up the frequency